requirements.txt
python-telegram-bot==20.7
anthropic==0.40.0
gspread==6.1.2
google-auth==2.29.0
APScheduler==3.10.4
pytz==2024.1
